// src/client/models/line.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Navitia API - SNCF
 * Complete API documentation for Navitia/SNCF public transportation API. Based on https://doc.navitia.io
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { CommercialMode } from './commercial-mode';
// May contain unused imports in some cases
// @ts-ignore
import type { Link } from './link';
// May contain unused imports in some cases
// @ts-ignore
import type { Network } from './network';
// May contain unused imports in some cases
// @ts-ignore
import type { PhysicalMode } from './physical-mode';

export interface Line {
    /**
     * Line ID
     */
    'id'?: string;
    /**
     * Line name
     */
    'name'?: string;
    /**
     * Line code
     */
    'code'?: string;
    /**
     * Line color (hex)
     */
    'color'?: string;
    /**
     * Text color (hex)
     */
    'text_color'?: string;
    'commercial_mode'?: CommercialMode;
    'physical_modes'?: Array<PhysicalMode>;
    'network'?: Network;
    'links'?: Array<Link>;
}